ACalcBrain:  - fix all values priority no int but NSInteger - popOperandOfStack return NSdictionary (isNeedShowGrad string, NSNumber result) - gradArrayFromNumber new function to return grad array

+++ ITSCalc/ACalcBrain.h
@@ -19,7 +19,7 @@
 -(BOOL) deleteLastElement;
 
 
--(void) getResultAndPutAsArguByPriotiy:(int)priority; //may be its need to be hiden???
+-(void) getResultAndPutAsArguByPriotiy:(NSInteger)priority; //may be its need to be hiden???
 -(id) count;
 -(id) perfomOperation:(NSString *) operation;
 -(void) insertBracket: (BOOL) isRigthOpen;
